requirements.txt
apify_client
newsplease
newspaper4k
openai
pika
python-dateutil
python-dotenv
pytz
requests
tldextract
pymongo
langdetect
openpyxl
